Address review: real behavior tests, cover the upload path and auth helpers
- refresh_access_token's tests now use a real temp .env file (monkeypatched
  ENV_PATH) instead of mocking set_key, asserting on the file's actual
  content rather than just that set_key was called with plausible args.
- Add build_authorize_url/extract_code/exchange_authorization_code tests
  in strava/auth.py — these were pulled out into standalone functions by
  the restructure specifically for reuse/testability, but had none.
  exchange_authorization_code's test uses the same real-.env-file approach
  as refresh_access_token.
- Add two tests exercising cli/sync.py's main() real upload path in-process
  (Strava mocked via `responses`, a real temp .env, monkeypatched argv) —
  the only path previously under test was --dry-run, so refresh -> upload
  -> poll -> mute being wired together correctly, the printed activity
  URL, and the --public flag's effect on muting were all untested.
- Add a structural test over the whole EXERCISE_TYPE_MAP (every value is
  well-formed upper-snake-case), not just the 7 entries spot-checked by
  example — guards the exact bug class the map's own comments describe
  recurring, across all ~40 entries instead of the tested handful.
- Fix pyright errors in test_strava_client.py (responses' `.body` is typed
  BodyType | None, so `in`/`.split()` needed an isinstance narrowing) and
  widen poll_upload's timeout_s to float, matching how it's used and
  tested (0.05s).

+++ liftostrava/src/liftostrava/strava/client.py
@@ -44,7 +44,7 @@ def upload_activity(access_token: str, workout: Workout, sport_type: str) -> dic
     return resp.json()
 
 
-def poll_upload(access_token: str, upload_id: int, timeout_s: int = 30) -> dict:
+def poll_upload(access_token: str, upload_id: int, timeout_s: float = 30) -> dict:
     # Starts unset rather than only being assigned inside the loop: with a
     # very small timeout_s the loop body can run zero times, and the
     # TimeoutError below used to reference this while still unbound,
